Make Munin node optional

+++ docker-cmd.sh
@@ -28,8 +28,14 @@ sudo -u munin -- /usr/sbin/rrdcached \
   -m 0660 -l unix:/run/munin/rrdcached.sock \
   -w 1800 -z 1800 -f 3600
 
-# Configure munin node in the docker
-munin-node-configure --shell --suggest 2>/dev/null | bash
+if [ "$DISABLE_MUNIN_NODE" != "true" ]; then
+
+  # Configure munin node in the docker
+  munin-node-configure --shell --suggest 2>/dev/null | bash
+  
+  $NODES="localhost:127.0.0.1 $NODES"
+  
+fi
 
 # Generate node list
 [[ ! -z "$NODES" ]] && for NODE in $NODES
@@ -73,11 +79,15 @@ EOF
   munin-node-configure --shell --snmp "$HOST" --snmpcommunity "$COMMUNITY" | bash
 done
 
-# Remove plugins that doesn't work in docker
-rm /etc/munin/plugins/{cpuspeed,open_files,users,swap,proc_pri}
+if [ "$DISABLE_MUNIN_NODE" != "true" ]; then
 
-# Start munin node on the docker
-munin-node
+  # Remove plugins that doesn't work in docker
+  rm /etc/munin/plugins/{cpuspeed,open_files,users,swap,proc_pri}
+
+  # Start munin node on the docker
+  munin-node
+  
+fi
 
 # Run once before we start fcgi
 sudo -u munin -- /usr/bin/munin-cron munin
